+++ src/messages/Message.hpp
@@ -119,6 +119,8 @@ struct Message {
 
     std::shared_ptr<ChannelPointReward> reward = nullptr;
 
+    uint32_t bits{0};
+
     QJsonObject toJson() const;
 
     void freeze() const